+++ sdk/src/client/node_api/core/routes.rs
@@ -6,6 +6,7 @@
 use std::str::FromStr;
 
 use packable::PackableExt;
+use serde::{Deserialize, Serialize};
 use url::Url;
 
 use crate::{