+++ sdk/src/types/block/address/bech32.rs
@@ -1,6 +1,7 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+use alloc::string::String;
 use core::str::FromStr;
 
 use bech32::{FromBase32, ToBase32, Variant};