+++ sdk/CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `FilterOptions::{alias_ids, foundry_ids, nft_ids}` fields;
-- `Account::{alias_output, foundry_output, nft_output}` methods;
+- `Account::{unspent_alias_output, unspent_foundry_output, unspent_nft_output}` methods;
 
 ## 0.3.0 - 2023-05-02