+++ sdk/CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `FilterOptions::{alias_ids, foundry_ids, nft_ids}` fields;
 - `Account::{unspent_alias_output, unspent_foundry_output, unspent_nft_output}` methods;
-- `StrongholdAdapter::get_stronghold` method;
+- `StrongholdAdapter::inner` method;
 
 ### Removed